requirements.txt
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
fastapi==0.111.0
uvicorn==0.30.0
mlflow==2.12.1
numpy==1.26.4
pytest==8.2.2
setuptools==68.2.2
wheel>=0.42.0
